+++ packages/react-ui-kit/src/Theme/Sizes.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2022 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,10 +17,11 @@
  *
  */
 
-import {MemberData} from './MemberData';
-
-export interface Members {
-  hasMore: boolean;
-  members: MemberData[];
-  pagingState?: string;
-}
+export const fontSizes = {
+  extraSmall: '0.6875rem',
+  small: '0.75rem',
+  medium: '0.875rem',
+  base: '1rem',
+  large: '1.25rem',
+  extraLarge: '1.5rem',
+};